Extract CalculatePrice method from OrderDetailsController into a separate class

+++ WholesaleBeer.API/Models/DTO/AddBeerRequestDto.cs
@@ -9,10 +9,10 @@ namespace WholesaleBeer.API.Models.DTO
         public string Name { get; set; }
         [Required]
         [Range(0, 100, ErrorMessage = "Alcohol content percentage must be between 0 and 100")]
-        public float AlcoholContentPercentage { get; set; }
+        public double AlcoholContentPercentage { get; set; }
         [Required]
-        [Range(0, float.MaxValue, ErrorMessage = "Price must be greater than 0")]
-        public float Price { get; set; }
+        [Range(0, double.MaxValue, ErrorMessage = "Price must be greater than 0")]
+        public double Price { get; set; }
         [Required]
         public Guid BreweryId { get; set; }
     }
