+++ WholesaleBeer.API/Models/DTO/BeerDto.cs
@@ -7,8 +7,8 @@ namespace WholesaleBeer.API.Models.DTO
     {
         public Guid Id { get; set; }
         public string Name { get; set; }
-        public float AlcoholContentPercentage { get; set; }
-        public float Price { get; set; }
+        public double AlcoholContentPercentage { get; set; }
+        public double Price { get; set; }
 
         public Brewery Brewery { get; set; }
     }
